Correction du modele MVC

+++ Modele/DAO/JoueurDao.php
@@ -1,15 +1,18 @@
 <?php
 require_once 'ModeleDao.php';
-require_once 'Joueur.php';
+require_once __DIR__ . '/../Joueur.php';
 
+// Classe DAO contenant toutes la logique de la table Joueur
 class JoueurDao implements ModeleDao {
 
     private PDO $pdo;
 
+    // Contructeur principale
     public function __construct(PDO $pdo){
         $this->pdo = $pdo;
     }
 
+    // Récupère tous les joueurs de la bd
     public function getAll(): array{
         $sql = "SELECT * FROM joueur ORDER BY nom";
         $stmt = $this->pdo->query($sql);
@@ -31,6 +34,7 @@ class JoueurDao implements ModeleDao {
         return $joueurs;
     }
 
+    // Récupère un joueur par son id
     public function getById(int $id): ?Joueur{
         $sql = "SELECT * FROM joueur where idJoueur = :id ";
         $stmt = $this->pdo->prepare($sql);
@@ -54,6 +58,7 @@ class JoueurDao implements ModeleDao {
         );
     }
 
+    // Ajoute un joueur à la liste
     public function add(object $obj): bool{
         if (!($obj instanceof Joueur)) return false;
         $sql = "INSERT INTO joueur
@@ -71,6 +76,7 @@ class JoueurDao implements ModeleDao {
         ]);
     }
 
+    // met à jour un joueur parmi la liste
     public function update(object $obj): bool{
         if(!$obj instanceof Joueur) return false;
 
@@ -98,11 +104,37 @@ class JoueurDao implements ModeleDao {
         
     }
 
-    public function delete(int $id): bool{
+    // supprime un joueur parmi la liste
+    public function delete(object $obj): bool{
+        if(!$obj instanceof Joueur) return false;
         $sql = "DELETE FROM joueur WHERE idJoueur = :id";
         $stmt = $this->pdo->prepare($sql);
 
-        return $stmt->execute([':id' => $id]);
+        return $stmt->execute([':id' => $obj->getIdJoueur()]);
     }
+
+    // Recupere un joueur par son numéro de licence
+    public function getByNumLicence(string $numLicence): ?Joueur { // ?Joueur pour indiquer aue l'objet peut etre null
+        $sql = "SELECT * FROM joueur WHERE numLicence = :numLicence";
+        $stmt = $this->pdo->prepare($sql);
+        $stmt->bindValue(':numLicence', $numLicence);
+        $stmt->execute();
+        
+        $row = $stmt->fetch(PDO::FETCH_ASSOC);
+        
+        if (!$row) return null;
+        
+        return new Joueur(
+            $row['idJoueur'],
+            $row['numLicence'],
+            $row['nom'],
+            $row['prenom'],
+            $row['dateNaissance'],
+            $row['taille'],
+            $row['poids'],
+            $row['statut']
+        );
+    }
+
 }
 ?>
